Remove UserProfileDetails component
